The store functions were moved to separate files respective to their function.
